Modify basket; add class AdvancedBasket

+++ smart_rm/basket.py
@@ -25,7 +25,10 @@ from remove import (
 from error import OtherOSError
 from datetime import datetime
 from hashlib import sha256
-from logging import error
+from logging import (
+    error,
+    warning
+)
 # from subprocess import call
 
 
@@ -33,18 +36,15 @@ class Basket(object):
     def __init__(
         self,
         basket_location=expanduser("~/.local/share/basket"),
-        dry_run=False,
-        same_name_politics=""
     ):
         self.basket_location = basket_location
         self.file = join(self.basket_location, "files")
         self.info = join(self.basket_location, "info")
 
-        self.dry_run = dry_run
-        self.same_name_politics = same_name_politics
-
         self._trashinfo_config = ConfigParser()
         self._trashinfo_config.add_section("Trash Info")
+
+        self._move=""
 
     def move_to_basket(self, path):         # XXX fix dry_run
         path = abspath(path)
@@ -53,14 +53,8 @@ class Basket(object):
         self._check_cycle(path)
         self._check_empty_space()
 
-        if not self.dry_run:
+        if self._move(self.basket_location, path):
             self._make_trash_info_file(path)
-            try:
-                error(join(self.file, basename(path)))
-                error(path)
-                rename(path, join(self.file, basename(path)))
-            except OSError as why:
-                raise OtherOSError(why.errno, why.strerror, why.filename)
 
     def _check_cycle(self, path):
         prefix = commonprefix([path, self.file])
@@ -79,29 +73,29 @@ class Basket(object):
         except OSError as why:
             raise OtherOSError(why.errno, why.strerror, why.filename)
 
-    def _check_name_conflicts(self, path):
-        pass
-
     def _check_empty_space(self):
         pass
 
-    def restore_files(self, files_to_restore):
-        for file in files_to_restore:
-            file_path_in_files = join(self.file, file)
-            file_path_in_info = join(self.info, file + ".trashinfo")
+    def restore_file(self, file):
+        file_path_in_files = join(self.file, file)
+        file_path_in_info = join(self.info, file + ".trashinfo")
 
-            path_to_restore = self._get_path_from_trashinfo_files(
-                file_path_in_info
-            )       # XXX check basket?
-            move(file_path_in_files, path_to_restore)
-            remove_file(file_path_in_info)
+        path_to_restore = self._get_path_from_trashinfo_files(
+            file_path_in_info
+        )       # XXX check basket?
+
+        # self._check_basket()          # XXX similar to move code
+        # self._check_cycle(path)
+        # self._check_empty_space()
+        if self._restore(file_path_in_files, path_to_restore):
+            remove_file(file_path_in_info)      # else some warning
 
     def _get_path_from_trashinfo_files(self, trashinfo_file):
         self._trashinfo_config.read(trashinfo_file)
 
         return self._trashinfo_config.get("Trash Info", "Path")
 
-    def clear_basket(files_location, info_location):        # TODOOO@
+    def clear_basket(files_location, info_location):
         remove_directory_content(files_location)
         remove_directory_content(info_location)
 
@@ -121,15 +115,78 @@ class Basket(object):
             self._trashinfo_config.write(fp)
 
 
-# need mode to choose rename or not
-def rewrite_samename_files_without_asking(path, destination):
-    target_path = join(dirname(destination)), basename(path)
+class AdvancedBasket(object):
+    def __init__(
+        self,
+        basket_location=expanduser("~/.local/share/basket"),
+        solve_name_conflicts_when_move_to_basket="",
+        solve_name_conflicts_when_restore="",
+        clean_basket_politic="",
+        check_hash=False,
+    ):
+
+        self.basket = Basket(
+            basket_location=basket_location,
+            check_hash=check_hash
+        )
+
+    def clean(self):
+        pass
+
+    def move_file_to_basket(self, file):
+        pass
+
+    def restore_files(self, paths):
+        pass
+
+    def view_content(self):
+        self.basket.view_content()
+
+
+# need mode to choose rename or not; what to do with dry_run?
+def rewrite_samename_files_without_asking(path, destination, dry_run=False):
+    target_path = join(dirname(destination), basename(path))
     remove_tree(target_path)
     move(path, target_path)
+    return target_path
 
 
+# need message!
 def rewrite_samename_files_with_asking(path, destination):
+    answer = raw_input(
+        "Do you want to rewrite {0} in {1}".format(basename(path), destination)
+    )
+    if answer:
+        return rewrite_samename_files_without_asking(path, destination)
+    else:
+        raise_exception_if_samename_files(path, destination)
+
+
+def raise_exception_if_samename_files(path, destination):
+    raise OtherOSError(
+        "Already exists {0} in {1}".format(basename(path), destination)
+    )
+
+
+def ask_new_name_for_movable_object(path, destination):
     new_name = raw_input("Enter new name for {0}:".format(path))
+    new_path = join(destination, new_name)
 
     if new_name:
-        move(path, join(destination, new_name))
+        move(path, new_path)
+        return new_path
+    else:
+        raise_exception_if_samename_files(path, destination)
+
+
+def new_name_for_movable_object_depending_on_amount_of_samename_files(
+    path,
+    destination
+):
+    count_of_samename_files = listdir(destination).count(basename(path))
+    new_path = join(
+        destination,
+        basename(path) + ".{0}".format(count_of_samename_files)
+    )
+    move(path, new_path)
+    return new_path
